Change prototype of API function iio_buffer_push
It now returns a ssize_t value instead of an int value, which is more
consistent with iio_buffer_refill and also makes sense, as
iio_buffer_push uses iio_device_write_raw which returns a ssize_t
value.

+++ buffer.c
@@ -71,7 +71,7 @@ ssize_t iio_buffer_refill(struct iio_buffer *buffer)
 	return read;
 }
 
-int iio_buffer_push(const struct iio_buffer *buffer)
+ssize_t iio_buffer_push(const struct iio_buffer *buffer)
 {
 	if (!buffer->is_output)
 		return -EINVAL;
